fix: move funtion to lib folder
