+++ src/components/Reactions/QuickEmojiReactions/BaseQuickEmojiReactions.js
@@ -8,8 +8,10 @@ import AddReactionBubble from '../AddReactionBubble';
 import CONST from '../../../CONST';
 import styles from '../../../styles/styles';
 import ONYXKEYS from '../../../ONYXKEYS';
-import getPreferredEmojiCode from '../getPreferredEmojiCode';
 import Tooltip from '../../Tooltip';
+import * as EmojiUtils from '../../../libs/EmojiUtils';
+
+
 
 const baseQuickEmojiReactionsPropTypes = {
     /**
@@ -55,7 +57,7 @@ const BaseQuickEmojiReactions = (props) => (
                 focusable={false}
             >
                 <EmojiReactionBubble
-                    emojiCodes={[getPreferredEmojiCode(emoji, props.preferredSkinTone)]}
+                    emojiCodes={[EmojiUtils.getPreferredEmojiCode(emoji, props.preferredSkinTone)]}
                     isContextMenu
                     onPress={() => {
                         props.onEmojiSelected(emoji);
